@@ -29,7 +29,7 @@
     "react": "16.13.0",
     "react-dom": "16.13.0",
     "react-jsx-parser": "1.28.1",
-    "webpack": "^5.74.0",
+    "webpack": "^5.76.1",
     "webpack-cli": "^4.10.0"
   },
   "peerDependencies": {
