@@ -31,7 +31,7 @@
     "ts-jest": "^29.0.3",
     "ts-loader": "^9.4.1",
     "typescript": "^4.8.4",
-    "webpack": "^5.74.0",
+    "webpack": "^5.76.1",
     "webpack-cli": "^4.10.0"
   },
   "peerDependencies": {
