@@ -1,6 +1,6 @@
 {
   "name": "dash-renderer",
-  "version": "1.14.5",
+  "version": "1.15.0",
   "description": "render dash components in react",
   "main": "build/dash_renderer.min.js",
   "scripts": {
@@ -77,7 +77,7 @@
     "style-loader": "^3.3.1",
     "ts-loader": "^9.4.1",
     "typescript": "^4.8.4",
-    "webpack": "^5.74.0",
+    "webpack": "^5.76.2",
     "webpack-cli": "^4.10.0",
     "whatwg-fetch": "^3.6.2"
   },
