@@ -31,7 +31,8 @@ function generateElements(graphs, profile, extraLinks) {
     const elements = [];
     const structure = {};
 
-    function recordNode(id, property) {
+    function recordNode(id, rawProperty) {
+        const property = rawProperty.split('@')[0];
         const idStr = stringifyId(id);
         const idType = typeof id === 'object' ? 'wildcard' : 'component';
 
@@ -155,6 +156,20 @@ function flattenInputs(inArray, final) {
         }
     });
     return final;
+}
+
+function cleanOutputId(outputId) {
+    return outputId
+        .replace(/(^\.\.|\.\.$)/g, '')
+        .split('...')
+        .reduce(
+            (agg, next) =>
+                agg.concat(
+                    next.replace(/(.*\..*)(@.+)$/, (a, b) => b + ' (Duplicate)')
+                ),
+            []
+        )
+        .join('...');
 }
 
 // len('__dash_callback__.')
@@ -326,7 +341,7 @@ function CallbackGraph() {
 
                 // Remove uid and set profile.
                 const callbackOutputId = data.id.slice(cbPrefixLen);
-                elementName = callbackOutputId.replace(/(^\.\.|\.\.$)/g, '');
+                elementName = cleanOutputId(callbackOutputId);
                 const cbProfile = profile.callbacks[callbackOutputId];
                 if (cbProfile) {
                     const {
