@@ -127,7 +127,10 @@ const observer: IStoreObserverDefinition<IStoreState> = {
                     );
 
                     const basePath = getPath(oldPaths, parsedId);
-                    const oldObj = path(getPath(oldPaths, parsedId), oldLayout);
+                    if (!basePath) {
+                        return;
+                    }
+                    const oldObj = path(basePath, oldLayout);
 
                     const childrenProps = pathOr(
                         'defaultValue',
@@ -138,7 +141,8 @@ const observer: IStoreObserverDefinition<IStoreState> = {
                     const handlePaths = (
                         children: any,
                         oldChildren: any,
-                        oldChildrenPath: any[]
+                        oldChildrenPath: any[],
+                        filterRoot: any = false
                     ) => {
                         const oPaths = getState().paths;
                         const paths = computePaths(
@@ -152,7 +156,8 @@ const observer: IStoreObserverDefinition<IStoreState> = {
                         requestedCallbacks = concat(
                             requestedCallbacks,
                             getLayoutCallbacks(graphs, paths, children, {
-                                chunkPath: oldChildrenPath
+                                chunkPath: oldChildrenPath,
+                                filterRoot
                             }).map(rcb => ({
                                 ...rcb,
                                 predecessors
@@ -166,7 +171,8 @@ const observer: IStoreObserverDefinition<IStoreState> = {
                             getLayoutCallbacks(graphs, oldPaths, oldChildren, {
                                 removedArrayInputsOnly: true,
                                 newPaths: paths,
-                                chunkPath: oldChildrenPath
+                                chunkPath: oldChildrenPath,
+                                filterRoot
                             }).map(rcb => ({
                                 ...rcb,
                                 predecessors
@@ -204,7 +210,8 @@ const observer: IStoreObserverDefinition<IStoreState> = {
                                     }
                                 },
                                 oldObj,
-                                basePath
+                                basePath,
+                                keys(appliedProps)
                             );
                             // Only do it once for the component.
                             recomputed = true;
