@@ -53,9 +53,10 @@ def until_not(
 
 
 class contains_text:
-    def __init__(self, selector, text):
+    def __init__(self, selector, text, timeout):
         self.selector = selector
         self.text = text
+        self.timeout = timeout
 
     def __call__(self, driver):
         try:
@@ -66,6 +67,17 @@ class contains_text:
             )
         except WebDriverException:
             return False
+
+    def message(self, driver):
+        try:
+            element = self._get_element(driver)
+            text = "found: " + str(element.text) or str(element.get_attribute("value"))
+        except WebDriverException:
+            text = f"{self.selector} not found"
+        return f"text -> {self.text} not found inside element within {self.timeout}s, {text}"
+
+    def _get_element(self, driver):
+        return driver.find_element(By.CSS_SELECTOR, self.selector)
 
 
 class contains_class:
@@ -86,13 +98,14 @@ class contains_class:
 
 
 class text_to_equal:
-    def __init__(self, selector, text):
+    def __init__(self, selector, text, timeout):
         self.selector = selector
         self.text = text
+        self.timeout = timeout
 
     def __call__(self, driver):
         try:
-            elem = driver.find_element(By.CSS_SELECTOR, self.selector)
+            elem = self._get_element(driver)
             logger.debug("text to equal {%s} => expected %s", elem.text, self.text)
             return (
                 str(elem.text) == self.text
@@ -100,6 +113,17 @@ class text_to_equal:
             )
         except WebDriverException:
             return False
+
+    def message(self, driver):
+        try:
+            element = self._get_element(driver)
+            text = "found: " + str(element.text) or str(element.get_attribute("value"))
+        except WebDriverException:
+            text = f"{self.selector} not found"
+        return f"text -> {self.text} not found within {self.timeout}s, {text}"
+
+    def _get_element(self, driver):
+        return driver.find_element(By.CSS_SELECTOR, self.selector)
 
 
 class style_to_equal:
